Add create-react-app item (#121)
Hey!

This PR adds a new item to the open-source list. It differs from the other create-react-app items in the list as it adds `TailwindCSS` to the react-scripts itself. This way it is even possible to add `PurgeCSS` for production builds.

To create a new project:

```bash
npx create-react-app --scripts-version=react-scripts-tailwindcss my-app
```